Refactoring to maintain consistency in formats
